+++ app/data/spells/earth.ts
@@ -1,0 +1,106 @@
+import { Spell } from "../../types/spell.ts";
+
+export const earthSpell: Array<Spell> = [
+  {
+    id: "52",
+    name: "Shield",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "53",
+    name: "Slow",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "54",
+    name: "Stone Skin",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "55",
+    name: "View Earth",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "56",
+    name: " Death Ripple",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "57",
+    name: "Quicksand",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "58",
+    name: "Animate Dead",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "59",
+    name: "Anti-Magic",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "60",
+    name: "Earthquake",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "61",
+    name: "Force Field",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "62",
+    name: "Protection from Earth",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "63",
+    name: "Meteor Shower",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "64",
+    name: "Resurrection",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "65",
+    name: "Sorrow",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "66",
+    name: "Town Portal",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "67",
+    name: "Implosion",
+    level: 5,
+    effect: "",
+  },
+  {
+    id: "68",
+    name: "Summon Earth Elemental",
+    level: 5,
+    effect: "",
+  },
+];